Add utils/nerf.py
